Replace remaining Unity 6.5 InstanceID APIs in MCPForUnity
Editor and leftover Runtime lookups now use GetEntityId plus EntityId.ToULong/FromULong and EntityIdToObject. Find calls drop obsolete sort-mode overloads. JSON instanceID fields stay named the same.

+++ Assets/MCPForUnity/Editor/Tools/FindGameObjects.cs
@@ -59,7 +59,7 @@ namespace MCPForUnity.Editor.Tools
                 var allIds = GameObjectLookup.SearchGameObjects(searchMethod, searchTerm, includeInactive, 0);
                 
                 // Use standard pagination response
-                var paginatedResult = PaginationResponse<int>.Create(allIds, pagination);
+                var paginatedResult = PaginationResponse<ulong>.Create(allIds, pagination);
 
                 return new SuccessResponse("Found GameObjects", new
                 {
